DescribedEntityInfoController replaced with NameDescriptionInfoView
